Remove javascript from the dashboard
Earlier javascript was being used to generate
the recipe search URI dynamically based on selected
drop down.

Since the app is supposed to work without javascript
support, that javascript was removed and instead an
intermediate end point was added which takes in necessary
information from the form as query parameters.

And uses those query parameters to construct the URI
to redirect to (this is the construction that javascript
was doing earlier on the client side)